move fees logic to its own module
