+++ runtime/common/src/lib.rs
@@ -15,12 +15,10 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
-#[cfg(test)]
-mod tests;
-
 pub mod account_conversion;
 pub mod apis;
 pub mod evm;
+pub mod fees;
 pub mod gateway;
 pub mod migrations;
 pub mod oracle;
@@ -88,92 +86,6 @@ pub mod xcm_fees {
 
 	pub fn dollar(decimals: u32) -> Balance {
 		10u128.saturating_pow(decimals)
-	}
-}
-
-pub mod fees {
-	use cfg_primitives::{
-		constants::{CENTI_CFG, TREASURY_FEE_RATIO},
-		types::Balance,
-	};
-	use frame_support::{
-		traits::{Currency, Imbalance, OnUnbalanced},
-		weights::{
-			constants::ExtrinsicBaseWeight, WeightToFeeCoefficient, WeightToFeeCoefficients,
-			WeightToFeePolynomial,
-		},
-	};
-	use smallvec::smallvec;
-	use sp_arithmetic::Perbill;
-
-	pub type NegativeImbalance<R> = <pallet_balances::Pallet<R> as Currency<
-		<R as frame_system::Config>::AccountId,
-	>>::NegativeImbalance;
-
-	struct ToAuthor<R>(sp_std::marker::PhantomData<R>);
-	impl<R> OnUnbalanced<NegativeImbalance<R>> for ToAuthor<R>
-	where
-		R: pallet_balances::Config + pallet_authorship::Config,
-	{
-		fn on_nonzero_unbalanced(amount: NegativeImbalance<R>) {
-			if let Some(author) = <pallet_authorship::Pallet<R>>::author() {
-				<pallet_balances::Pallet<R>>::resolve_creating(&author, amount);
-			}
-		}
-	}
-
-	pub struct DealWithFees<R>(sp_std::marker::PhantomData<R>);
-	impl<R> OnUnbalanced<NegativeImbalance<R>> for DealWithFees<R>
-	where
-		R: pallet_balances::Config + pallet_treasury::Config + pallet_authorship::Config,
-		pallet_treasury::Pallet<R>: OnUnbalanced<NegativeImbalance<R>>,
-	{
-		fn on_unbalanceds<B>(mut fees_then_tips: impl Iterator<Item = NegativeImbalance<R>>) {
-			if let Some(fees) = fees_then_tips.next() {
-				// for fees, split the destination
-				let (treasury_amount, mut author_amount) = fees.ration(
-					TREASURY_FEE_RATIO.deconstruct(),
-					(Perbill::one() - TREASURY_FEE_RATIO).deconstruct(),
-				);
-				if let Some(tips) = fees_then_tips.next() {
-					// for tips, if any, 100% to author
-					tips.merge_into(&mut author_amount);
-				}
-
-				use pallet_treasury::Pallet as Treasury;
-				<Treasury<R> as OnUnbalanced<_>>::on_unbalanced(treasury_amount);
-				<ToAuthor<R> as OnUnbalanced<_>>::on_unbalanced(author_amount);
-			}
-		}
-	}
-
-	/// Handles converting a weight scalar to a fee value, based on the scale
-	/// and granularity of the node's balance type.
-	///
-	/// This should typically create a mapping between the following ranges:
-	///   - [0, frame_system::MaximumBlockWeight]
-	///   - [Balance::min, Balance::max]
-	///
-	/// Yet, it can be used for any other sort of change to weight-fee. Some
-	/// examples being:
-	///   - Setting it to `0` will essentially disable the weight fee.
-	///   - Setting it to `1` will cause the literal `#[weight = x]` values to
-	///     be charged.
-	pub struct WeightToFee;
-	impl WeightToFeePolynomial for WeightToFee {
-		type Balance = Balance;
-
-		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-			let p = CENTI_CFG;
-			let q = 10 * Balance::from(ExtrinsicBaseWeight::get().ref_time());
-
-			smallvec!(WeightToFeeCoefficient {
-				degree: 1,
-				negative: false,
-				coeff_frac: Perbill::from_rational(p % q, q),
-				coeff_integer: p / q,
-			})
-		}
 	}
 }
